+++ src/main/java/uz/dckroff/statisfy/dto/analytics/ActivityCountDTO.java
@@ -1,0 +1,15 @@
+package uz.dckroff.statisfy.dto.analytics;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ActivityCountDTO {
+    private String name;
+    private Long count;
+}